fix: Consolidate Bitcoin module structure and update import paths
