+++ src/custom/bitcoin_agent/analytics/__init__.py
@@ -4,4 +4,4 @@ Bitcoin Analytics Module
 This module provides analytics and visualization capabilities for Bitcoin data.
 """
 
-from src.custom.bitcoin.analytics.price_chart import generate_price_chart, generate_price_history_chart 
+from src.custom.bitcoin_agent.analytics.price_chart import generate_price_chart, generate_price_history_chart
