+++ src/custom/bitcoin_agent/integration/__init__.py
@@ -4,7 +4,7 @@ Bitcoin Integration Module
 Provides direct interaction with the Bitcoin blockchain through APIs.
 """
 
-from src.custom.bitcoin.integration.BitcoinIntegration import (
+from src.custom.bitcoin_agent.integration.BitcoinIntegration import (
     BitcoinIntegration,
     BitcoinIntegrationConfiguration,
     as_tools
