+++ src/custom/bitcoin_agent/models/__init__.py
@@ -4,4 +4,4 @@ Bitcoin Models Module
 This module provides model selection capabilities for Bitcoin agents and analytics.
 """
 
-from src.custom.bitcoin.models.model_provider import get_model, ModelProvider, ModelConfig 
+from src.custom.bitcoin_agent.models.model_provider import get_model, ModelProvider, ModelConfig
