+++ src/custom/bitcoin_agent/pipeline/__init__.py
@@ -4,7 +4,7 @@ Bitcoin Pipeline Module
 Contains pipelines for processing Bitcoin transaction data and mapping it to ontology structures.
 """
 
-from src.custom.bitcoin.pipeline.BitcoinTransactionPipeline import (
+from src.custom.bitcoin_agent.pipeline.BitcoinTransactionPipeline import (
     BitcoinTransactionPipeline,
     BitcoinTransactionPipelineConfiguration,
     BitcoinTransactionPipelineParameters
