+++ src/custom/bitcoin_agent/workflow/ChatBitcoinAgentWorkflow.py
@@ -1,7 +1,7 @@
 from abi.workflow import Workflow, WorkflowConfiguration
 from abi.workflow.workflow import WorkflowParameters
-from src.custom.bitcoin.integration import BitcoinIntegration, BitcoinIntegrationConfiguration
-from src.custom.bitcoin.pipeline import BitcoinTransactionPipeline, BitcoinTransactionPipelineConfiguration
+from src.custom.bitcoin_agent.integration import BitcoinIntegration, BitcoinIntegrationConfiguration
+from src.custom.bitcoin_agent.pipeline import BitcoinTransactionPipeline, BitcoinTransactionPipelineConfiguration
 from src import secret, config
 from dataclasses import dataclass
 from pydantic import BaseModel, Field
